Beginnings of the framework
packages are for organizing classes of different purpose.
api - Mod API for other mods to interface with. contains the main API utility class, aswell as ICurrencies and CurrenciesCapabilitie.
ICurrencies is the Capability Interface to create a Currency-holding Capability for various objects.

capability - classes relevant to Capability Implementations within the mod. such as the ICurrencies implementation for players, PlayerCurrencyCapability.
We may want to ship capabilities for other objects such as Block/TileEntities, in order to permit "Bank Blocks" or other such things out of the box.

command - classes relevant to the mod's Commands.
Currently contains setbalance (sets your balance for the given currency) and listcurrencies (lists debug info of all registered currencies)

data - classes relevant to Data-driven information, such as the CurrencyType objects used in the registry, and the registry listing in EternalCurrenciesRegistries, and the data loader in CurrencyTypeManager.