enabled test directory and added FTBQuestRequirement test directory lets stuff be tested without having to disable it before compile, will be useful.
FTBQuestRequirement - provide the ID of a quest as a Hex string of the "quest" field in the requirement, and it'll require that quest to be completed to satisfy the requirement. probably anyway, shop is too broken to test.

added some temporary assets from malcolmriley/unused_textures(https://github.com/malcolmriley/unused-textures)

ItemDebitCard shows a different tooltip line when unlinked to cue you in on its use.

ItemDebitCard and ItemCurrencyCheque use messages now appear on the Actionbar instead of game chat.

updated version

unfortunately this doesn't run on servers right now, because some calls to EternalCurrenciesAPI#getRegisteredCurrencies and getCurrencyData are serverside calls without RegistryAccess provided, and currently idk where to fetch a level or anything from statically for a Dedicated Server, might not be possible and require some reworks.